feat: button picture from tools to customization
