+++ src/client/locales/en-US/translation.json
@@ -6,7 +6,6 @@
     "duplicate": "Duplicate",
     "hideTextboxes": "Hide textboxes",
     "showTextboxes": "Show textboxes",
-    "addImage": "Add an image",
     "undo": "Undo",
     "redo": "Redo",
     "eraseAll": "Erase all",
@@ -50,6 +49,7 @@
     "text": "Text",
     "customization": "Customization",
     "addText": "Add Text",
+    "addImage": "Add an image",
     "color": "Color",
     "textAlign": "Horizontal align",
     "alignVertical": "Vertical Align",
