+++ src/client/locales/fr/translation.json
@@ -6,7 +6,6 @@
     "duplicate": "Dupliquer",
     "hideTextboxes": "Cacher les zones de texte",
     "showTextboxes": "Afficher les zones de texte",
-    "addImage": "Ajouter une image",
     "undo": "Annuler",
     "redo": "Rétablir",
     "eraseAll": "Écraser tout",
@@ -50,6 +49,7 @@
     "text": "Texte",
     "customization": "Personnalisation",
     "addText": "Ajouter un texte",
+    "addImage": "Ajouter une image",
     "color": "Couleur",
     "textAlign": "Alignement horizontal",
     "alignVertical": "Alignement vertical",
