+++ src/client/scss/styles.scss
@@ -173,7 +173,7 @@ body {
         padding: 17px 18px;
       }
 
-      .add-text-button {
+      .add-item-button {
         margin-top: auto;
       }
     }
