+++ src/client/ts/components/Aside/Tabs/Customization/customization.scss
@@ -87,7 +87,7 @@
       }
     }
 
-    .add-text-button {
+    .add-item-button {
       border-radius: 0;
       border: 0;
       padding: 0;
@@ -97,11 +97,32 @@
       cursor: pointer;
       outline: none;
       background-color: var(--blue-primary);
-      color: #ffffff;
-      font-weight: 500;
       flex-shrink: 0;
-      box-shadow: 0px 2px 4px -1px rgba(0,0,0,0.75);
-      height: 48px;
+      height: 36px;
+      position: relative;
+      color: #ffffff;
+      font-size: 0.77rem;
+
+      & + .add-item-button {
+        border-top: 1px solid rgba(255,255,255,0.15);
+        box-shadow: 0px 2px 4px -1px rgba(0,0,0,0.65);
+        margin-bottom: 8px;
+      }
+
+      input {
+        position: absolute;
+        z-index: -1;
+        opacity: 0;
+        width: 0;
+        height: 0;
+      }
+
+      span {
+        display: inline;
+        line-height: 1;
+        font-weight: 500;
+      }
+
       &:hover {
         background-color: var(--blue-secondary);
       }
