+++ src/client/ts/components/Tools/Tools.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { useRef, RefObject, useContext, useCallback } from 'react'
+import { useRef, RefObject, useContext } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useTranslation } from 'react-i18next'
 import { useHistory } from 'react-router-dom'
@@ -10,58 +10,24 @@ import {
   ERASE_ALL,
   RESET,
   TOGGLE_EXPORT_MODAL,
-  TOGGLE_THEME,
-  ADD_ITEM
+  TOGGLE_THEME
 } from '@client/store/reducer/constants'
-import { IMAGE_ADDED } from '../../shared/constants'
 import Faq from '@client/components/Modal/Faq/Faq'
 import { DefaultContext, DefaultState, DefaultDispatch } from '@client/store/DefaultContext'
 import { useWindowWidth } from '@client/ts/shared/hooks'
 import { EditorContext, EditorInt, EditorDispatch } from '@client/store/EditorContext'
-import { toBase64, endWithExt } from '@client/utils/index'
 import './tools.scss'
 
 const Tools = (): JSX.Element => {
   const history = useHistory()
   const faqModal: RefObject<any> = useRef(null)
-  const uploadInput: RefObject<HTMLInputElement> = useRef(null)
   const { t } = useTranslation()
   const { isMinLgSize } = useWindowWidth()
   const [{ theme }, dispatch]: [DefaultState, DefaultDispatch] = useContext(DefaultContext)
-  const [{ showTextAreas, memeSelected, canUndo, canRedo, canErazeAll, saveToEditor }, dispatchEditor]: [
+  const [{ showTextAreas, memeSelected, canUndo, canRedo, canErazeAll }, dispatchEditor]: [
     EditorInt,
     EditorDispatch
   ] = useContext(EditorContext)
-
-  const handleUploadImagebox = useCallback(
-    async (e: React.ChangeEvent<HTMLInputElement>): Promise<void> => {
-      const files = e.currentTarget.files
-
-      if (files.length > 1 || !endWithExt(['.jpg', '.png', 'jpeg'], files[0].name)) {
-        uploadInput.current.value = ''
-        return
-      }
-
-      const file = files[0]
-
-      try {
-        const img = await toBase64(file)
-        img.name = file.name
-
-        saveToEditor({
-          type: ADD_ITEM,
-          historyType: IMAGE_ADDED,
-          itemType: 'image',
-          img
-        })
-      } catch (error) {
-        console.error(error)
-      } finally {
-        uploadInput.current.value = ''
-      }
-    },
-    [saveToEditor]
-  )
 
   return (
     <div className="tools">
@@ -116,29 +82,6 @@ const Tools = (): JSX.Element => {
             <FontAwesomeIcon icon={['fas', 'eraser']} />
           </button>
         </li>
-        {isMinLgSize && (
-          <li>
-            <label htmlFor="upload-imagebox">
-              <span
-                role="button"
-                aria-label={t('attr.addImage')}
-                className={`tools-list-btn ${!memeSelected ? 'tools-list-btn-disabled' : ''}`}
-                data-tooltip={t('attr.addImage')}
-              >
-                <FontAwesomeIcon icon={['fas', 'chart-area']} />
-              </span>
-              <input
-                type="file"
-                ref={uploadInput}
-                disabled={!memeSelected}
-                onChange={handleUploadImagebox}
-                className="upload-imagebox-input"
-                accept="image/png, image/jpeg"
-                id="upload-imagebox"
-              />
-            </label>
-          </li>
-        )}
         <li>
           <button
             aria-label={t('attr.reset')}
